main_test.go
package main

import (
	"net/http"
	"net/http/httptest"
	"testing"
)

func TestHome(t *testing.T) {
	req, err := http.NewRequest("GET", "/", nil)
	if err != nil {
		t.Fatal(err)
	}
	rr := httptest.NewRecorder()
	handler := http.HandlerFunc(Home)
	handler.ServeHTTP(rr, req)
	if status := rr.Code; status != http.StatusOK {
		t.Errorf("handler returned wrong status code: got %v want %v",
			status, http.StatusOK)
	}

	var statusCode = rr.Code

	// Check the response body is what we expect.
	expected := 200
	if statusCode != expected {
		t.Errorf("response result: got %v want %v",
			rr.Code, expected)
	}
}
func TestContact(t *testing.T) {
	req, err := http.NewRequest("GET", "/contact", nil)
	if err != nil {
		t.Fatal(err)
	}
	rr := httptest.NewRecorder()
	handler := http.HandlerFunc(Contact)
	handler.ServeHTTP(rr, req)
	if status := rr.Code; status != http.StatusOK {
		t.Errorf("handler returned wrong status code: got %v want %v",
			status, http.StatusOK)
	}
	var statusCode = rr.Code

	// Check the response body is what we expect.
	expected := 200
	if statusCode != expected {
		t.Errorf("response result: got %v want %v",
			rr.Code, expected)
	}
}